requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.1
django-admin-sortable2==2.2.3
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
setuptools==75.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.44.0
whitenoise==6.7.0
